Move a few functions into the driver core
xf86Wacom will eventually be the X interface only, move functions that
aren't a direct interface for the X driver out of there.
